+++ packages/terra-collapsible-menu-view/src/CollapsibleMenuViewItem.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import Button from 'terra-button';
 import ButtonGroup from 'terra-button-group';
-import 'terra-base/lib/baseStyles';
 import Menu from './_CollapsibleMenu';
 import styles from './CollapsibleMenuView.module.scss';