+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -1,8 +1,10 @@
+
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReactDatePicker from 'react-datepicker';
-import 'terra-base/lib/baseStyles';
 import ResponsiveElement from 'terra-responsive-element';
+
+/* eslint-disable-next-line  */
+import ReactDatePicker from './react-datepicker';
 import PopperContainer from './_PopperContainer';
 import DateInput from './DateInput';
 import DateUtil from './DateUtil';
@@ -71,6 +73,10 @@ const propTypes = {
    */
   requestFocus: PropTypes.func,
   /**
+   * Whether or not the date is required.
+   */
+  required: PropTypes.bool,
+  /**
    * An ISO 8601 string representation of the initial value to show in the date input.
    * This prop name is derived from react-datepicker but is analogous to value in a form input field.
    */
@@ -90,6 +96,7 @@ const defaultProps = {
   onClickOutside: undefined,
   onSelect: undefined,
   releaseFocus: undefined,
+  required: false,
   requestFocus: undefined,
   selectedDate: undefined,
 };
@@ -239,6 +246,7 @@ class DatePicker extends React.Component {
       onClickOutside,
       onSelect,
       requestFocus,
+      required,
       releaseFocus,
       selectedDate,
       ...customProps
@@ -267,6 +275,7 @@ class DatePicker extends React.Component {
         onChangeRaw={this.handleChangeRaw}
         onClickOutside={this.handleOnClickOutside}
         onSelect={this.handleOnSelect}
+        required={required}
         customInput={(
           <DateInput
             onInputFocus={this.handleOnInputFocus}
@@ -304,6 +313,7 @@ class DatePicker extends React.Component {
         onChangeRaw={this.handleChangeRaw}
         onClickOutside={this.handleOnClickOutside}
         onSelect={this.handleOnSelect}
+        required={required}
         customInput={(
           <DateInput
             onInputFocus={this.handleOnInputFocus}