+++ packages/terra-dialog-modal/src/DialogModal.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import AbstractModal from 'terra-abstract-modal';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import 'terra-base/lib/baseStyles';
 import styles from './DialogModal.module.scss';
 
 const cx = classNames.bind(styles);