+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+3.11.0 - (March 26, 2019)
+------------------
+### Changed
+* Updated prop onChange call logic with respect to ambiguous date-times.
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
+
 3.10.0 - (March 20, 2019)
 ------------------
 ### Changed